// src/enums/RecordGroupColor.ts
export enum RecordGroupColor {
    RED = '#F15B50',
    ORANGE = '#F7943D',
    YELLOW = '#F9D03B',
    GREEN = '#59C36A',
    MINT = '#66CDB9',
    SKY_BLUE = '#5EAED6',
    LIGHT_BLUE = '#64B9F3',
    BLUE = '#4B7CF5',
    PURPLE = '#6B62E8',
    DEEP_PURPLE = '#A364E7',
    BROWN = '#8B7355',
    GRAY = '#808080'
} 